Move king checkers to separate module
